refactor: folder structure for request refund page
