+++ apps/web/src/pages/job/[jobId]/actor/[actorId]/refund.tsx
@@ -1,1 +1,1 @@
-export { default } from 'modules/job/refund'
+export { default } from 'modules/job/refund/pages'
